restructured code based on review
