+++ abods-api/package.json
@@ -43,6 +43,7 @@
     "graphql-scalars": "^1.24.2",
     "graphql-tag": "^2.12.6",
     "haversine-distance": "^1.2.3",
+    "jest-mock-extended": "^4.0.0",
     "kysely": "^0.27.4",
     "pg": "^8.13.1",
     "pino": "^9.1.0",
@@ -66,7 +67,7 @@
     "serverless-dotenv-plugin": "^6.0.0",
     "serverless-offline": "^13.9.0",
     "serverless-plugin-typescript": "^2.1.5",
-    "ts-jest": "^29.4.3",
+    "ts-jest": "^29.4.4",
     "typescript": "^5.6.3",
     "typescript-eslint": "^8.11.0"
   },
