+++ abods-api/src/lib/otp.test.ts
@@ -1,4 +1,4 @@
-import { getKyselyFiltersForOTPQuery } from "../../../src/resolvers/otpFunctions";
+import { getKyselyFiltersForOTPQuery, kyselyFilterForAdminIds } from "./otp";
 import {
   Dialect,
   DummyDriver,
@@ -7,8 +7,8 @@ import {
   PostgresIntrospector,
   PostgresQueryCompiler,
 } from "kysely";
-import { DB } from "../../../src/kysely";
-import { MatchType, Granularity } from "../../../src/types/generated";
+import { DB } from "../kysely";
+import { MatchType } from "../types/generated";
 
 const dummyDialect: Dialect = {
   createDriver: () => new DummyDriver(),
@@ -19,6 +19,33 @@ const dummyDialect: Dialect = {
 
 const dummyKysely = new Kysely<DB>({
   dialect: dummyDialect,
+});
+
+describe("kyselyFilterForAdminIds", () => {
+  it("adds admin area filter when adminAreaIds are provided", () => {
+    const mockQuery = dummyKysely.selectFrom("timetable_summary_operator_t");
+
+    const adminAreaIds = ["1", "2", "3"];
+    const result = kyselyFilterForAdminIds(mockQuery as never, adminAreaIds);
+
+    const compiledQuery = result.compile();
+
+    expect(compiledQuery.sql).toContain(
+      "admin_areas && ARRAY[$1, $2, $3]::int4[]",
+    );
+    expect(compiledQuery.parameters).toEqual(["1", "2", "3"]);
+  });
+
+  it("returns original query when adminAreaIds is empty", () => {
+    const mockQuery = dummyKysely.selectFrom("timetable_summary_operator_t");
+
+    const adminAreaIds: string[] = [];
+    const result = kyselyFilterForAdminIds(mockQuery as never, adminAreaIds);
+
+    const compiledQuery = result.compile();
+    expect(compiledQuery.sql).not.toContain("admin_areas");
+    expect(compiledQuery.parameters).toEqual([]);
+  });
 });
 
 describe("getKyselyFiltersForOTPQuery", () => {
