+++ abods-api/src/resolvers/otp/onTimePerformance.ts
@@ -1,59 +1,36 @@
 import {
-  AdminAreasType,
   DelayFrequencyType,
   Direction,
-  FrequentServiceInfoInputType,
-  FrequentServiceInfoType,
-  FrequentServiceType,
   Granularity,
-  HeadwayInputType,
-  HeadwayMetricsTypeResolvers,
-  HeadwayOverviewType,
-  HeadwayTimeSeriesType,
-  LineType,
   MatchType,
   Maybe,
   OnTimePerformanceTypeResolvers,
   OperatorPerformancePage,
   OperatorPerformanceType,
-  OperatorType,
-  PerformanceInputType,
   PunctualityDayOfWeekType,
   PunctualityTimeOfDayType,
   PunctualityTimeSeriesType,
   PunctualityTotalsType,
-  QueryResolvers,
   RankingOrder,
   Resolvers,
-  ServiceInfoType,
-  ServicePatternType,
   ServicePerformanceType,
   ServicePunctualityType,
   StopPerformanceType,
-} from "../types/generated.js";
-import logger from "../logger.js";
-import {
-  getFormattedDate,
-  toUkTime,
-  userSelectedDateAsUtc,
-} from "../lib/dayjs.js";
+} from "../../types/generated.js";
+import logger from "../../logger.js";
+import { getFormattedDate, userSelectedDateAsUtc } from "../../lib/dayjs.js";
 import {
   compareThresholds,
-  getFrequentServiceActualHours,
+  getKyselyFiltersForOTPQuery,
   getOperatorsForUser,
-  getSummaryStopsTotalHours,
-} from "../lib/otp.js";
-import { getDayOfWeekNumbers } from "../lib/utils.js";
-import { emptyResolver, requireUserSession } from "./helpers.js";
-import {
-  getUserOperatorIds,
-  getUserOperatorIdsQuery,
-} from "../lib/operators.js";
-import { Kysely, sql } from "kysely";
-import { DB } from "../kysely.js";
-import { listServiceLinks } from "../lib/common.js";
+  kyselyFilterForAdminIds,
+} from "../../lib/otp.js";
+import { requireUserSession } from "../helpers.js";
+import { getUserOperatorIds } from "../../lib/operators.js";
+import { sql } from "kysely";
 import dayjs from "dayjs";
-import { executeQuery } from "../lib/kysely.js";
+import { executeQuery } from "../../lib/dbKysely.js";
+import { OTPSummaryTables } from "../../types/extra.js";
 
 interface DayCount {
   dayOfWeek: number;
@@ -61,215 +38,6 @@ interface DayCount {
   onTime: number;
   late: number;
 }
-
-export const getOperatorList: QueryResolvers["operators"] = async (
-  _,
-  args,
-  context,
-): Promise<OperatorType[]> => {
-  const user = await requireUserSession(context);
-
-  let query = context.kysely
-    .selectFrom("service_details as s")
-    .where(
-      "s.operator_noc",
-      "in",
-      getUserOperatorIdsQuery(context.kysely, user),
-    )
-    .innerJoin("all_operators as a", "a.operatorref", "s.operator_noc")
-    .innerJoin(
-      "noc_adminarea as n",
-      "n.national_operator_code",
-      "s.operator_noc",
-    );
-  if (args.filterBy?.operatorIds && args.filterBy.operatorIds.length > 0) {
-    query = query.where("s.operator_noc", "in", args.filterBy.operatorIds);
-  }
-
-  if (args.filterBy?.orgId) {
-    query = query
-      .innerJoin(
-        "bods_organisationoperator as boo",
-        "boo.operatorref",
-        "a.operatorref",
-      )
-      .where("boo.organisation_id", "=", args.filterBy?.orgId);
-  }
-  return await query
-    .groupBy(["a.name", "s.operator_noc"])
-    .select((eb) => [
-      eb.fn.coalesce("name", sql.lit("<unknown>")).as("name"),
-      eb.fn.coalesce("operator_noc", sql.lit("<unknown>")).as("operatorId"),
-      sql<string>`string_agg(distinct n.adminarea_id::text, ',')`.as(
-        "adminAreaIds",
-      ),
-    ])
-    .orderBy("name")
-    .execute()
-    .then((x) =>
-      x.map((o) => ({
-        name: o.name,
-        operatorId: o.operatorId,
-        nocCode: o.operatorId,
-        adminAreaIds: o.adminAreaIds.split(","),
-      })),
-    );
-};
-
-export const getServiceInfo: QueryResolvers["serviceInfo"] = async (
-  _,
-  args,
-  context,
-): Promise<Maybe<ServiceInfoType>> => {
-  const user = await requireUserSession(context);
-  try {
-    const userOperatorIds = await getUserOperatorIds(user, context.kysely);
-    const service = await context.db.expected_services.findFirst({
-      where: {
-        noc_and_line_and_servicecode: args.serviceId,
-      },
-      select: {
-        operator_noc: true,
-        line_name: true,
-        service_name: true,
-      },
-    });
-
-    if (!service) {
-      throw Error("No service found");
-    }
-
-    if (userOperatorIds.includes(service.operator_noc)) {
-      return {
-        serviceId: args.serviceId,
-        serviceNumber: service.line_name,
-        serviceName: service.service_name,
-      };
-    } else throw Error("User does not have access to service");
-  } catch (error) {
-    logger.error(error, "An error occurred when getting service info");
-    return null;
-  }
-};
-
-export const getLines: QueryResolvers["lines"] = async (
-  _,
-  args,
-  context,
-): Promise<LineType[]> => {
-  const user = await requireUserSession(context);
-
-  if (args.operatorIds.length === 0) return [];
-
-  let query = context.kysely
-    .selectFrom("expected_services")
-    .where("operator_noc", "in", getUserOperatorIdsQuery(context.kysely, user))
-    .where("operator_noc", "in", args.operatorIds);
-
-  const inputDate = userSelectedDateAsUtc(args.inputDate).toDate();
-  if (args.endDate) {
-    const endDate = userSelectedDateAsUtc(args.endDate).toDate();
-    query = query
-      .where("date_of_journey", ">=", inputDate)
-      .where("date_of_journey", "<", endDate);
-  } else {
-    query = query.where("date_of_journey", "=", inputDate);
-  }
-
-  return query
-    .select("noc_and_line_and_servicecode as id")
-    .select("service_name as name")
-    .select("line_name as number")
-    .select("admin_area_id as adminAreaIds")
-    .distinctOn("noc_and_line_and_servicecode")
-    .execute();
-};
-
-async function getOtpServiceLinks(
-  stops: { stopId: string; lon: number; stopName: string; lat: number }[],
-  stopIdList: string[],
-  db: Kysely<DB>,
-) {
-  stops = stops.sort(
-    (a, b) => stopIdList.indexOf(a.stopId) - stopIdList.indexOf(b.stopId),
-  );
-  return listServiceLinks(stops, db);
-}
-
-export const getServicePatterns: QueryResolvers["servicePatterns"] = async (
-  _,
-  args,
-  context,
-): Promise<ServicePatternType[]> => {
-  await requireUserSession(context);
-  const routesQueryResults = await context.db.distinct_routes.findMany({
-    where: {
-      servicepattern_route: {
-        noc_and_line_and_servicecode: args.lineId,
-      },
-    },
-    select: {
-      id: true,
-      route: true,
-    },
-  });
-  const routes = routesQueryResults.map((n) => ({
-    ...n,
-    stopIds: n.route.split(","),
-  }));
-  const allStopIds = [...new Set(routes.flatMap((n) => n.stopIds))];
-  const stopQueryResults = await context.db.naptan_stoppoint_latlong.findMany({
-    where: {
-      atco_code: { in: allStopIds },
-      NOT: {
-        atco_code: null,
-        longitude: null,
-        latitude: null,
-      },
-    },
-    select: {
-      common_name: true,
-      atco_code: true,
-      longitude: true,
-      latitude: true,
-    },
-  });
-  const stopDetails = stopQueryResults.map((n) => ({
-    stopName: n.common_name,
-    // workaround for nullable db columns that can probably be not null, the where clause should exclude null for now
-    stopId: n.atco_code!,
-    lon: n.longitude!,
-    lat: n.latitude!,
-  }));
-
-  const result: ServicePatternType[] = [];
-  for (const route of routes) {
-    const stops = stopDetails.filter((s) => route.stopIds.includes(s.stopId));
-    const serviceLinks = await getOtpServiceLinks(
-      stops,
-      route.stopIds,
-      context.kysely,
-    );
-    result.push({
-      stops,
-      servicePatternId: route.id.toString(),
-      serviceLinks,
-    });
-  }
-  return result;
-};
-
-export const kyselyFilterForAdminIds = (
-  query: ReturnType<typeof getKyselyFiltersForOTPQuery>,
-  adminAreaIds: string[],
-) => {
-  if (adminAreaIds && adminAreaIds.length > 0) {
-    query = query.where(
-      sql<boolean>`admin_areas && ARRAY[${sql.join(adminAreaIds)}]::int4[]`,
-    );
-  }
-  return query;
-};
 
 export const getPunctualityOverview: OnTimePerformanceTypeResolvers["punctualityOverview"] =
   async (_, args, context): Promise<Maybe<PunctualityTotalsType>> => {
@@ -1416,381 +1184,7 @@ export const getServicePerformance: OnTimePerformanceTypeResolvers["servicePerfo
     }
   };
 
-// -> OPERATOR PAGE
-export const getFrequentServices: HeadwayMetricsTypeResolvers["frequentServices"] =
-  async (_, args, context): Promise<Maybe<FrequentServiceType[]>> => {
-    const user = await requireUserSession(context);
-    try {
-      const userOperatorIds = await getUserOperatorIds(user, context.kysely);
-      if (userOperatorIds.includes(args.operatorId)) {
-        const results =
-          await context.db.timetable_frequent_summary_services.findMany({
-            where: {
-              operator_noc: args.operatorId,
-              date_of_journey: {
-                gte: userSelectedDateAsUtc(args.fromTimestamp).toDate(),
-                lt: userSelectedDateAsUtc(args.toTimestamp).toDate(),
-              },
-            },
-            select: {
-              noc_and_line_and_servicecode: true,
-            },
-            distinct: ["noc_and_line_and_servicecode"],
-          });
-
-        return results.map((result) => ({
-          serviceId: result.noc_and_line_and_servicecode,
-        }));
-      }
-
-      return [];
-    } catch (error) {
-      logger.error(error, "An error occurred when getting frequent services");
-      return null;
-    }
-  };
-
-export const getFrequentServiceInfo: HeadwayMetricsTypeResolvers["frequentServiceInfo"] =
-  async (_, args, context): Promise<Maybe<FrequentServiceInfoType>> => {
-    const user = await requireUserSession(context);
-    try {
-      const userOperatorIds = await getUserOperatorIds(user, context.kysely);
-
-      const [totalHours, actualHours] = await Promise.all([
-        getSummaryStopsTotalHours(context.kysely, args.inputs, userOperatorIds),
-        getFrequentServiceActualHours(
-          context.kysely,
-          args.inputs,
-          userOperatorIds,
-        ),
-      ]);
-
-      return {
-        numHours: actualHours,
-        totalHours: totalHours,
-      };
-    } catch (error) {
-      logger.error(
-        error,
-        "An error occurred when getting frequent service info",
-      );
-      return null;
-    }
-  };
-
-export const getHeadwayOverview: HeadwayMetricsTypeResolvers["headwayOverview"] =
-  async (_, args, context): Promise<Maybe<HeadwayOverviewType>> => {
-    const user = await requireUserSession(context);
-    try {
-      const { filters } = args.inputs;
-      const { startTime, endTime } = filters || {};
-
-      const userOperatorIds = await getUserOperatorIds(user, context.kysely);
-      const summarySubQuery = getKyselyFiltersForOTPQuery(
-        context.kysely,
-        "timetable_frequent_summary_services",
-        args.inputs,
-        userOperatorIds,
-      );
-
-      // Needs to be aliased separately. Need to find out why
-      const aliasedSubQuery = summarySubQuery
-        .where("headway_stops_count", ">", sql.lit("0"))
-        .where("excess_wait_time", "is not", null)
-        .as("summary");
-
-      let mainQuery = context.kysely
-        .selectFrom(aliasedSubQuery)
-        .select(["headway_stops_count", "excess_wait_time"]);
-
-      if (startTime || endTime) {
-        const start = Number((startTime ?? "00:00").split(":")[0]);
-        const end = Number((endTime ?? "23:59").split(":")[0]);
-        mainQuery = mainQuery
-          .where("hour", ">=", start)
-          .where("hour", "<=", end);
-      }
-
-      const results = await executeQuery(mainQuery);
-
-      if (results.length < 1) {
-        return {
-          excess: undefined,
-        };
-      }
-
-      let headway = {
-        excessWaitTime: 0,
-        headwayCount: 0,
-      };
-
-      headway = results.reduce((acc, currentHeadway) => {
-        acc.excessWaitTime +=
-          Number(currentHeadway.excess_wait_time) *
-          Number(currentHeadway.headway_stops_count);
-        acc.headwayCount += Number(currentHeadway.headway_stops_count);
-
-        return acc;
-      }, headway);
-
-      return {
-        excess: headway.excessWaitTime / headway.headwayCount,
-      };
-    } catch (error) {
-      logger.error(error, "An error occurred when getting headway overview");
-      return null;
-    }
-  };
-
-export const getHeadwayTimeSeries: HeadwayMetricsTypeResolvers["headwayTimeSeries"] =
-  async (_, args, context): Promise<Maybe<HeadwayTimeSeriesType[]>> => {
-    const user = await requireUserSession(context);
-    try {
-      const { filters } = args.inputs;
-      const { startTime, endTime } = filters || {};
-
-      const userOperatorIds = await getUserOperatorIds(user, context.kysely);
-      const summarySubQuery = getKyselyFiltersForOTPQuery(
-        context.kysely,
-        "timetable_frequent_summary_services",
-        args.inputs,
-        userOperatorIds,
-      );
-
-      // Needs to be aliased separately. Need to find out why
-      const aliasedSubQuery = summarySubQuery
-        .where("headway_stops_count", ">", sql.lit("0"))
-        .where("actual_headway", "is not", null)
-        .where("expected_headway", "is not", null)
-        .where("excess_wait_time", "is not", null)
-        .as("summary");
-
-      let mainQuery = context.kysely
-        .selectFrom(aliasedSubQuery)
-        .select([
-          "date_of_journey",
-          "departure_hour",
-          "headway_stops_count",
-          "actual_headway",
-          "expected_headway",
-          "excess_wait_time",
-        ]);
-
-      if (startTime || endTime) {
-        const start = Number((startTime ?? "00:00").split(":")[0]);
-        const end = Number((endTime ?? "23:59").split(":")[0]);
-        mainQuery = mainQuery
-          .where("hour", ">=", start)
-          .where("hour", "<=", end);
-      }
-
-      const results = await executeQuery(mainQuery);
-
-      const headwayMap: Record<
-        string,
-        {
-          actual_headway: number;
-          expected_headway: number;
-          excess_wait_time: number;
-          headway_stops_count: number;
-        }
-      > = {};
-
-      results.map((result) => {
-        if (result.departure_hour) {
-          const time = toUkTime(result.departure_hour);
-          const timeIndex =
-            args.inputs.filters.granularity === Granularity.Day
-              ? time.startOf("day")
-              : time;
-          const index = timeIndex.toISOString();
-          const headwayData = (headwayMap[index] ??= {
-            actual_headway: 0,
-            expected_headway: 0,
-            excess_wait_time: 0,
-            headway_stops_count: 0,
-          });
-          headwayData.actual_headway +=
-            // We've filtered out null values in where clause so its fine to assert not null
-            Number(result.actual_headway) * Number(result.headway_stops_count);
-          headwayData.expected_headway +=
-            // We've filtered out null values in where clause so its fine to assert not null
-            Number(result.expected_headway) *
-            Number(result.headway_stops_count);
-          headwayData.excess_wait_time +=
-            // We've filtered out null values in where clause so its fine to assert not null
-            Number(result.excess_wait_time) *
-            Number(result.headway_stops_count);
-          headwayData.headway_stops_count += Number(result.headway_stops_count);
-        }
-      });
-
-      const returnHeadways: HeadwayTimeSeriesType[] = [];
-
-      for (const [departure_hour, headway] of Object.entries(headwayMap)) {
-        returnHeadways.push({
-          ts: new Date(departure_hour),
-          // Prevent confusion on the front end by rounding to the nearest second before converting to number of minutes
-          actual: headway.actual_headway / headway.headway_stops_count,
-          scheduled: headway.expected_headway / headway.headway_stops_count,
-          excess: headway.excess_wait_time / headway.headway_stops_count,
-        });
-      }
-
-      return returnHeadways.sort((a, b) => {
-        if (dayjs(a.ts).isBefore(dayjs(b.ts))) return -1;
-        return 1;
-      });
-    } catch (error) {
-      logger.error(error, "An error occurred when getting headway time series");
-      return null;
-    }
-  };
-
-export const getAdminAreas: QueryResolvers["adminAreas"] = async (
-  _,
-  __,
-  context,
-): Promise<Maybe<AdminAreasType[]>> => {
-  const user = await requireUserSession(context);
-  try {
-    const userOperatorIds = await getUserOperatorIds(user, context.kysely);
-    const adminAreaRecords = await context.db.noc_adminarea.findMany({
-      where: {
-        national_operator_code: {
-          in: userOperatorIds,
-        },
-      },
-      select: {
-        adminarea_id: true,
-      },
-    });
-
-    if (adminAreaRecords) {
-      const adminareaIds = adminAreaRecords.map((a) => a.adminarea_id);
-      const adminAreas = await context.db.naptan_adminarea_with_shape.findMany({
-        where: {
-          id: {
-            in: adminareaIds,
-          },
-        },
-      });
-
-      if (!adminAreas) {
-        throw Error("No admin areas found");
-      }
-
-      return adminAreas.map((adminArea) => ({
-        id: adminArea.id.toString(),
-        name: adminArea.name,
-        shape: adminArea.st_asgeojson,
-      }));
-    }
-
-    return null;
-  } catch (error) {
-    logger.error(error, "An error occurred when getting admin areas");
-    return null;
-  }
-};
-
-type OTPSummaryTables = keyof Pick<
-  DB,
-  | "timetable_summary_service_tz"
-  | "timetable_summary_operator_t"
-  | "timetable_summary_stops_tz"
-  | "timetable_frequent_summary_services"
-  | "timetable_threshold_summary"
->;
-
-export const getKyselyFiltersForOTPQuery = (
-  db: Kysely<DB>,
-  tableName: OTPSummaryTables,
-  inputs: PerformanceInputType &
-    HeadwayInputType &
-    FrequentServiceInfoInputType,
-  userOperatorNocList: string[],
-) => {
-  const { fromTimestamp, toTimestamp, filters } = inputs || {};
-  const {
-    timingPointsOnly,
-    operatorId,
-    maxDelay,
-    minDelay,
-    lineIds,
-    lineId,
-    dayOfWeekFlags,
-    matchType,
-  } = filters || {};
-  const operatorIds = filters?.operatorIds ?? [];
-
-  const nocListToFilter: string[] = userOperatorNocList
-    .filter((o) => operatorIds.includes(o))
-    .filter((o) => !operatorId || o === operatorId);
-
-  const startDateUtc = userSelectedDateAsUtc(fromTimestamp);
-  const endDateUtc = userSelectedDateAsUtc(toTimestamp);
-
-  // assign maxlate and maxearly filters (maxearly switched to positive for db condition)
-  const maxLateNumber = maxDelay ?? 0;
-  const maxEarlyNumber = minDelay ? Math.abs(minDelay) : 0;
-
-  const isServiceGranularity = lineIds && lineIds.length > 0;
-
-  const lines = lineId ? [lineId] : lineIds;
-
-  let query = db
-    .selectFrom(tableName)
-    .selectAll()
-    .select((eb) =>
-      sql<number>`EXTRACT(HOUR FROM ${eb.ref("departure_hour")} AT TIME ZONE 'Europe/London')`.as(
-        "hour",
-      ),
-    )
-    .where("date_of_journey", ">=", startDateUtc.toDate())
-    .where("date_of_journey", "<", endDateUtc.toDate());
-
-  if (nocListToFilter.length > 0) {
-    query = query.where("operator_noc", "in", nocListToFilter);
-  }
-  if (matchType && matchType === MatchType.Evidenced) {
-    query = query.where("estimated", "=", false);
-  }
-
-  if (timingPointsOnly) {
-    query = query.where("is_timing_point", "=", timingPointsOnly);
-  }
-
-  if (dayOfWeekFlags) {
-    const dayOfWeekNumbers = getDayOfWeekNumbers(dayOfWeekFlags);
-    query = query.where("day_of_week", "in", dayOfWeekNumbers);
-  }
-
-  if (maxEarlyNumber > 0 && !isServiceGranularity) {
-    query = query.where("max_early", "<=", maxEarlyNumber);
-  }
-
-  if (maxLateNumber > 0 && !isServiceGranularity) {
-    query = query.where("max_late", "<=", maxLateNumber);
-  }
-
-  if (lines) {
-    query = query.where("noc_and_line_and_servicecode", "in", lines);
-  }
-
-  return query;
-};
-
-const otpResolvers: Resolvers = {
-  Query: {
-    operators: getOperatorList,
-    onTimePerformance: emptyResolver,
-    headwayMetrics: emptyResolver,
-    serviceInfo: getServiceInfo,
-    adminAreas: getAdminAreas,
-    lines: getLines,
-    servicePatterns: getServicePatterns,
-  },
+const onTimePerformanceResolvers: Resolvers = {
   OnTimePerformanceType: {
     delayFrequency: getDelayFrequency,
     operatorPerformance: getOperatorPerformance,
@@ -1802,12 +1196,6 @@ const otpResolvers: Resolvers = {
     stopPerformance: getStopPerformance,
     servicePerformance: getServicePerformance,
   },
-  HeadwayMetricsType: {
-    frequentServices: getFrequentServices,
-    frequentServiceInfo: getFrequentServiceInfo,
-    headwayOverview: getHeadwayOverview,
-    headwayTimeSeries: getHeadwayTimeSeries,
-  },
 };
 
-export default otpResolvers;
+export default onTimePerformanceResolvers;
