+++ abods-api/src/types/extra.ts
@@ -1,9 +1,10 @@
 import express from "express";
 import { IncomingHttpHeaders } from "http";
-import { PrismaClient } from "@prisma/client";
+import { PrismaClient, Timetable } from "@prisma/client";
 import { Kysely } from "kysely";
 import { DB } from "../kysely.js";
 import { BaseContext } from "@apollo/server";
+import { CorridorStatsInputType } from "./generated.js";
 
 export interface AuthContext {
   allowedTokenHash: string;
@@ -27,3 +28,32 @@ export interface SessionUser {
   }[];
   isGlobalUser?: boolean;
 }
+
+export type OTPSummaryTables = keyof Pick<
+  DB,
+  | "timetable_summary_service_tz"
+  | "timetable_summary_operator_t"
+  | "timetable_summary_stops_tz"
+  | "timetable_frequent_summary_services"
+  | "timetable_threshold_summary"
+>;
+
+export interface StatsCache {
+  inputs: CorridorStatsInputType;
+  corridorTransits: TimetableType[][];
+}
+
+export type TimetableType = Pick<
+  Timetable,
+  | "atco_code"
+  | "stop_index"
+  | "actual_departure_time"
+  | "timestamp_after_estimate"
+  | "expected_departure_time"
+  | "operator_noc"
+  | "service_code"
+  | "line_name"
+  | "date_of_journey"
+  | "vehiclejourney_id"
+  | "group_id"
+>;
