--- abods-api/tests/lambda/apiinfo_query.json
@@ -1,16 +1,0 @@
-{
-    "version": "2",
-    "headers": {
-      "content-type": "application/json"
-    },
-    "isBase64Encoded": false,
-    "rawQueryString": "",
-    "requestContext": {
-      "http": {
-        "method": "POST"
-      }
-    },
-    "rawPath": "/",
-    "routeKey": "/",
-    "body": "{\"operationName\": null, \"variables\": null, \"query\": \"{   x { ID } }\"}"
-  }
