--- abods-api/tests/lambda/login_query.json
@@ -1,17 +1,0 @@
-{
-    "version": "2",
-    "headers": {
-      "content-type": "application/json"
-    },
-    "isBase64Encoded": false,
-    "rawQueryString": "",
-    "requestContext": {
-      "http": {
-        "method": "POST"
-      }
-    },
-    "rawPath": "/",
-    "routeKey": "/",
-    "body": "{\"operationName\":\"login\",\"variables\":{\"username\":\"emily.ostridge@kpmg.co.uk\",\"password\":\"123\"},\"query\":\"mutation login($username: String!, $password: String!) {login(username: $username, password: $password) { success expiresAt __typename }}\"}"
-    
-  }
