+++ abods-api/tsconfig.json
@@ -11,6 +11,6 @@
     "allowSyntheticDefaultImports": true,
     "strict": true
   },
-  "include": ["src/**/*"],
-  "exclude": ["**/*.spec.ts", "**/*.test.ts", "node_modules"]
+  "include": ["src/**/*", "src/**/*.test.ts"],
+  "exclude": ["**/*.spec.ts", "node_modules"]
 }
