+++ package.json
@@ -25,6 +25,7 @@
     "graphql-scalars": "^1.24.2",
     "husky": "^9.1.7",
     "lint-staged": "^16.1.6",
+    "node-mocks-http": "^1.17.2",
     "prettier": "3.2.5"
   },
   "lint-staged": {
